add file data mytraining
