+++ src/components/MyTrainingTableView/MyTrainingTableViewData.js
@@ -1,0 +1,145 @@
+import { Rate } from "antd";
+
+const data = [
+  {
+    key: "1",
+    name: "Angular JS",
+    eventType: "Online class",
+    eventPeriod: "08/18/2019, 11:53:32",
+    trainierName: "juwitha Susanti",
+    rating: <Rate disabled defaultValue={3} />,
+    adtionalInfo: "New York No. 1 Lake Park",
+  },
+  {
+    key: "2",
+    name: "Seminar Kit",
+    eventType: "Offline class",
+    eventPeriod: "06/18/2019, 13:53:32",
+    trainierName: "Zuki Yah",
+    rating: <Rate disabled defaultValue={2} />,
+    adtionalInfo: "London No. 1 Lake Park",
+  },
+  {
+    key: "3",
+    name: "Golang",
+    eventType: "Online class",
+    eventPeriod: "07/18/2019, 15:53:32",
+    trainierName: "Bambang Sopo",
+    rating: <Rate disabled defaultValue={1} />,
+    adtionalInfo: "Sidney No. 1 Lake Park",
+  },
+  {
+    key: "4",
+    name: "React JS",
+    eventType: "Offline class",
+    eventPeriod: "02/18/2019, 12:53:32",
+    trainierName: "Jojo Jiji",
+    rating: <Rate disabled defaultValue={5} />,
+    adtionalInfo: "London No. 2 Lake Park",
+  },
+  {
+    key: "5",
+    name: "Vue Js",
+    eventType: "Offline class",
+    eventPeriod: "03/10/2012, 12:53:32",
+    trainierName: "Tepos Sekale",
+    rating: <Rate disabled defaultValue={4} />,
+    adtionalInfo: "London No. 2 Lake Park",
+  },
+  {
+    key: "6",
+    name: "Front End",
+    eventType: "Online class",
+    eventPeriod: "12/29/2019, 12:53:32",
+    trainierName: "Naik ki",
+    rating: <Rate disabled defaultValue={5} />,
+    adtionalInfo: "London No. 2 Lake Park",
+  },
+  {
+    key: "7",
+    name: "Back End",
+    eventType: "Offline class",
+    eventPeriod: "11/17/2022, 12:53:32",
+    trainierName: "Adi sadidas ",
+    rating: <Rate disabled defaultValue={5} />,
+    adtionalInfo: "London No. 2 Lake Park",
+  },
+  {
+    key: "8",
+    name: "Full Stack ",
+    eventType: "Online class",
+    eventPeriod: "09/02/2022, 12:53:32",
+    trainierName: "Hero Pool",
+    rating: <Rate disabled defaultValue={5} />,
+    adtionalInfo: "London No. 2 Lake Park",
+  },
+  {
+    key: "9",
+    name: "Database",
+    eventType: "Offline class",
+    eventPeriod: "07/25/2022, 12:53:32",
+    trainierName: "Kiki kaka",
+    rating: <Rate disabled defaultValue={2} />,
+    adtionalInfo: "London No. 2 Lake Park",
+  },
+  {
+    key: "10",
+    name: "Mysql",
+    eventType: "Offline class",
+    eventPeriod: "05/12/2022, 12:53:32",
+    trainierName: "Lasono Hib",
+    rating: <Rate disabled defaultValue={5} />,
+    adtionalInfo: "London No. 2 Lake Park",
+  },
+  {
+    key: "11",
+    name: "Git",
+    eventType: "Offline class",
+    eventPeriod: "02/18/2019, 12:53:32",
+    trainierName: "Boy Wiliam",
+    rating: <Rate disabled defaultValue={0} />,
+    adtionalInfo: "London No. 2 Lake Park",
+  },
+];
+
+const columns = [
+  {
+    title: "Event Name",
+    dataIndex: "name",
+    key: "name",
+    render: (text) => <a>{text}</a>,
+    sorter: (a, b) => a.name.localeCompare(b.name),
+  },
+  {
+    title: "Event Type",
+    dataIndex: "eventType",
+    key: "eventType",
+    sorter: (a, b) => a.eventType.localeCompare(b.eventType),
+  },
+  {
+    title: "Event Period",
+    dataIndex: "eventPeriod",
+    key: "eventPeriod",
+    sorter: (a, b) => new Date(a.eventPeriod) - new Date(b.eventPeriod),
+  },
+  {
+    title: "Trainier Name",
+    dataIndex: "trainierName",
+    key: "trainierName",
+    sorter: (a, b) => a.trainierName.localeCompare(b.trainierName),
+  },
+  {
+    title: "Rating",
+    dataIndex: "rating",
+    key: "rating",
+    // sorter: (a, b) => a.defaultValue.localeCompare(b.defaultValue),
+  },
+  {
+    title: "Adtional Info",
+    dataIndex: "adtionalInfo",
+    key: "adtionalInfo",
+    sorter: (a, b) => a.adtionalInfo.localeCompare(b.adtionalInfo),
+  },
+];
+
+export { data, columns };
